+++ app/controllers/bookmarks_controller.rb
@@ -2,17 +2,21 @@ class BookmarksController < ApplicationController
   before_action :authenticate_user!
 
   def create
-    @product = Product.find(params[:id])
+    @product = Product.find(params[:product_id])
+    @variant = ProductVariant.find(params[:product_variant_id]) if params[:product_variant_id].present?
 
-    @bookmark = current_user.bookmarks.new(product: @product)
+    @bookmark = current_user.bookmarks.new(
+      product: @product,
+      product_variant: @variant || nil
+    )
 
     flash[:alert] = I18n.t(:generic_error_message) unless @bookmark.save
 
-    redirect_to product_path(id: @product.friendly_id)
+    redirect_back fallback_location: root_url
   end
 
   def destroy
-    @bookmark = current_user.bookmarks.find_by(id: params[:id])&.destroy
+    @bookmark = current_user.bookmarks.find(params[:id])&.destroy
 
     redirect_back fallback_location: root_url
   end