+++ db/migrate/20240308020714_add_variant_to_products_users_join_table.rb
@@ -1,0 +1,5 @@
+class AddVariantToProductsUsersJoinTable < ActiveRecord::Migration[7.1]
+  def change
+    add_column :products_users, :variant_id, :bigint
+  end
+end