+++ db/migrate/20240308030145_add_variant_id_to_bookmarks.rb
@@ -1,0 +1,6 @@
+class AddVariantIdToBookmarks < ActiveRecord::Migration[7.1]
+  def change
+    add_column :bookmarks, :variant_id, :bigint
+    add_column :prev_owneds, :variant_id, :bigint
+  end
+end