+++ db/migrate/20240308031958_rename_products_users_to_possessions.rb
@@ -1,0 +1,6 @@
+class RenameProductsUsersToPossessions < ActiveRecord::Migration[7.1]
+  def change
+    rename_table :products_users, :possessions
+    rename_column :possessions, :variant_id, :product_variant_id
+  end
+end