+++ db/migrate/20240308061938_add_id_to_setup_products.rb
@@ -1,0 +1,5 @@
+class AddIdToSetupProducts < ActiveRecord::Migration[7.1]
+  def change
+    add_column :setup_products, :id, :primary_key
+  end
+end