+++ db/migrate/20240308154248_allow_null_values_for_product_id_and_product_variant_id_in_setup_possessions.rb
@@ -1,0 +1,5 @@
+class AllowNullValuesForProductIdAndProductVariantIdInSetupPossessions < ActiveRecord::Migration[7.1]
+  def change
+    change_column :setup_possessions, :product_id, :bigint, null: true
+  end
+end