+++ db/migrate/20240309012049_add_index_to_sub_categories_on_slug.rb
@@ -1,0 +1,5 @@
+class AddIndexToSubCategoriesOnSlug < ActiveRecord::Migration[7.1]
+  def change
+    add_index :sub_categories, :slug, unique: true
+  end
+end