+++ db/migrate/20240311004026_add_discontinued_to_product_variant.rb
@@ -1,0 +1,8 @@
+class AddDiscontinuedToProductVariant < ActiveRecord::Migration[7.1]
+  def change
+    add_column :product_variants, :disabled, :boolean
+    add_column :product_variants, :discontinued_day, :integer
+    add_column :product_variants, :discontinued_month, :integer
+    add_column :product_variants, :discontinued_year, :integer
+  end
+end