+++ db/migrate/20240311004926_rename_product_variant_disabled_to_discontinued.rb
@@ -1,0 +1,5 @@
+class RenameProductVariantDisabledToDiscontinued < ActiveRecord::Migration[7.1]
+  def change
+    rename_column :product_variants, :disabled, :discontinued
+  end
+end